rename test_custom_data_sources.py to test_static_data_sources.py
