+++ corehq/apps/userreports/tests/test_static_data_sources.py
